Added temporary upgrades that can be selected upon reaching a new lvl Added lvl and experience system Added tanks textures Object spawn changed Allies can independently find targets and destroy them

+++ Tank Survivors Prototype/Assets/Scripts/Data/AliveEntityData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Alive Entity Data", menuName = "Alive Entity Data", order = 63)]
 public class AliveEntityData : ScriptableObject
 {
-    public int maxHealth;
+    public float maxHealth;
     public int startWeaponId;
 
     public float moveSpeed = 500;
